+++ hammer/prefabs.cpp
@@ -306,7 +306,7 @@ static int __cdecl SortPrefabs(CPrefab *a, CPrefab *b)
 //-----------------------------------------------------------------------------
 void CPrefabLibrary::Sort(void)
 {
-	int nPrefabs = Prefabs.GetCount();
+	intp nPrefabs = Prefabs.GetCount();
 	if (nPrefabs < 2)
 	{
 		return;
@@ -319,7 +319,7 @@ void CPrefabLibrary::Sort(void)
 	//
 	POSITION p = ENUM_START;
 	CPrefab *pPrefab = EnumPrefabs(p);
-	int iPrefab = 0;
+	intp iPrefab = 0;
 	while (pPrefab != NULL)
 	{
 		TmpPrefabArray[iPrefab++] = pPrefab;
@@ -335,7 +335,7 @@ void CPrefabLibrary::Sort(void)
 	// Store back in list in sorted order.
 	//
 	Prefabs.RemoveAll();
-	for (int i = 0; i < nPrefabs; i++)
+	for (intp i = 0; i < nPrefabs; i++)
 	{
 		Prefabs.AddTail(TmpPrefabArray[i]);
 	}
